refactor: moved isCLASSIFICATION functions to appropriate levels
