+++ app/models/representative-body.js
@@ -4,6 +4,7 @@ import {
   validateBelongsToOptional,
   validateHasManyOptional,
 } from '../validators/schema';
+import { RepresentativeOrganCodeList } from '../constants/Classification';
 
 export const BLACKLIST_RO = [
   'e224c637ba8bb0e5dfbb87da225b4652', // Executief van de Moslims van België
@@ -27,5 +28,9 @@ export default class RepresentativeBodyModel extends OrganizationModel {
       recognizedWorshipType: validateBelongsToOptional(),
       ministerPositions: validateHasManyOptional(),
     });
+  }
+
+  get isRepresentativeBody() {
+    return this._hasClassificationId(RepresentativeOrganCodeList);
   }
 }
