+++ app/models/worship-administrative-unit.js
@@ -5,7 +5,6 @@ import {
   validateRequiredWhenClassificationId,
 } from '../validators/schema';
 import { WorshipServiceCodeList } from '../constants/Classification';
-import { WITH_CENTRAL_WORSHIP_SERVICE } from './recognized-worship-type';
 
 export default class WorshipAdministrativeUnitModel extends AdministrativeUnitModel {
   @belongsTo('recognized-worship-type', {
@@ -41,14 +40,7 @@ export default class WorshipAdministrativeUnitModel extends AdministrativeUnitMo
     });
   }
 
-  get hasCentralWorshipService() {
-    return (
-      this.isWorshipService &&
-      this.#hasRecognizedWorshipTypeId(WITH_CENTRAL_WORSHIP_SERVICE)
-    );
-  }
-
-  #hasRecognizedWorshipTypeId(classificationIds) {
-    return classificationIds.includes(this.classification?.get('id'));
+  get isWorshipAdministrativeUnit() {
+    return this.isWorshipService || this.isCentralWorshipService;
   }
 }
